Move UniqueRoomList to seedu.address.model.room
