+++ src/test/java/seedu/address/model/room/UniqueRoomListTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model;
+package seedu.address.model.room;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -11,10 +11,6 @@ import java.util.List;
 import java.util.stream.IntStream;
 
 import org.junit.jupiter.api.Test;
-
-import seedu.address.model.room.Room;
-
-
 
 class UniqueRoomListTest {
